-converted um_eth_serv to use SG when talking on a socket. -um_eth_serv no longer supports AF_SOCKET or AF_NETLINK (to a tap device) -um_eth_tool comand line inteface changed -um_eth_tool can now specify a script that will be run when the   virtual inteface is admin'd up -um_eth_tool can specify which TAP should be used on the host
um_eth_serv is NOT needed to run ethernet based link between UMs or between
a UM and the host.  These cases can be handled by creating TAP interfaces
and bridging betweem them in differnt ways.

um_eth_serv persists so that non-ethernet interface (when the virtual drivers
are written) can use it to interconnect.